feature:WorkManager改为单例
